requirements.txt
numpy>=1.24
sounddevice>=0.4
pynput>=1.7; sys_platform != "darwin"
pyobjc-framework-Quartz>=10.0; sys_platform == "darwin"
